requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
requests>=2.28.0
google-generativeai>=0.7.0
PyMuPDF>=1.23.0,<2.0.0
Pillow>=10.0.0
youtube-transcript-api>=0.6.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
pydantic>=1.10.0,<3.0.0
streamlit>=1.32.0
PyPDF2>=3.0.0
PyYAML>=6.0
phidata>=2.4.0
pandas>=2.0.0
gunicorn>=21.0.0

